requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pymongo[srv]==4.7.3
python-dotenv==1.0.1
MarkupSafe==2.1.5
markdown==3.6
google-generativeai==0.8.3
